Minutes  Int Selection

+++ app/src/main/java/ru/fxy7ci/schf/TimerHolder.kt
@@ -2,5 +2,5 @@ package ru.fxy7ci.schf
 
 class TimerHolder (// Заданная температура
     var temperature: Byte = 20, // заданная температура
-    var timeDecMins: Byte = 10) {
+    var timeDecMins: Int = 10) {
 }
